MüsteriController ,Account Controller Düzenleme

+++ barber_management_system/Models/Hizmet.cs
@@ -6,7 +6,7 @@
         public string HizmetAd { get; set; }
         public decimal Fiyat { get; set; }
 
-        public int Dakika { get; set; } // Dakika cinsinden
+        public decimal Dakika { get; set; } // Dakika cinsinden
 
         // Hizmeti veren çalışanlar (Many-to-Many)
         public List<CalisanHizmet> calisanhizmetlist { get; set; }
